+++ src/check/pats/mod.rs
@@ -1,3 +1,6 @@
+pub use self::core::*;
 pub use self::mat::*;
 
+mod core;
+/// Match patterns.
 mod mat;